tests, histogram type hand tie resolving
